backend/requirements.txt
chromadb
langchain-community
langchain-google-genai
langchain-text-splitters
pymupdf
python-dotenv
fastapi
uvicorn
python-multipart
